# swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"version": "1.0",
		"title": "Wallet",
		"description": "description of the methods"
	},
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"servers": [
		{ "url": "https://wallet-backend.onrender.com/api" },
		{ "url": "http://localhost:7000/api" }
	],
...
	"paths": {
...
		"/transactions": {
			"post": {
				"tags": ["Transactions"],
				"summary": "Сreating a transaction",
				"parameters": [],
				"security": [{ "Bearer": [] }],
				"requestBody": {
					"description": "Registration's object",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionsRegistrationRequest"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TransactionsRegistrationResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request (Invalid request body)",
						"content": {}
					},
					"401 ": {
						"description": "Unauthorized",
						"content": {}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {}
					}
				}
			},
			"get": {
				"tags": ["Transactions"],
				"summary": "Get user's transactions",
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"type": "string",
						"description": "page of transaction",
						"example": "1"
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"type": "string",
						"description": "limit of transaction",
						"example": "5"
					}
				],
				"security": [{ "Bearer": [] }],

				"responses": {
					"201": {
						"description": "Successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetTransactionsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request (Invalid request body)",
						"content": {}
					},
					"401 ": {
						"description": "Unauthorized",
						"content": {}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"TransactionsRegistrationRequest": {
				"type": "object",
				"required": ["typeOperation", "category", "typeOperation"],
				"properties": {
					"typeOperation": {
						"type": "string",
						"enum": ["income", "expense"],
						"description": "Transactions type"
					},
					"category": {
						"type": "string",
						"description": "Transactions category"
					},
					"comment": {
						"type": "string",
						"description": "Transactions comment"
					},

					"amount": {
						"type": "number",
						"description": "Sum of transaction "
					},
					"date": {
						"type": "string",
						"description": "Date of transaction "
					}
				}
			},
			"TransactionsRegistrationResponse": {
				"type": "object",
				"properties": {
					"typeOperation": {
						"type": "string",
						"enum": ["income", "expense"],
						"description": "Transactions type"
					},
					"category": {
						"type": "string",
						"description": "Transactions category",
						"example": "Education"
					},
					"comment": {
						"type": "string",
						"description": "Transactions comment",
						"example": "Book"
					},
					"date": {
						"type": "string",
						"description": "Date of transaction ",
						"example": "Tue Nov 22 2022 12:58:30 GMT+0200"
					},
					"timestamps": {
						"type": "number",
						"description": "getTime() of transaction ",
						"example": "1669586400000"
					},

					"amount": {
						"type": "number",
						"description": "Sum of transaction ",
						"example": "100"
					},
					"balanceAfterTransaction": {
						"type": "string",
						"description": "Balance after transaction ",
						"example": "-50"
					},
					"owner": {
						"type": "string",
						"description": "Transaction's owner (id)",
						"example": "63837d687f0d6aafc551ffcc"
					},
					"createdAt": {
						"type": "string",
						"description": "Transaction's creating date",
						"example": "2022-11-27T15:08:24.572Z"
					}
				}
			},
			"GetTransactionsResponse": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string",
						"description": "Transactions id",
						"example": "6383803a9e42a3a435da9708"
					},
					"typeOperation": {
						"type": "string",
						"enum": ["income", "expense"],
						"description": "Transactions type"
					},
					"category": {
						"type": "string",
						"description": "Transactions category",
						"example": "Education"
					},
					"comment": {
						"type": "string",
						"description": "Transactions comment",
						"example": "Book"
					},
					"date": {
						"type": "string",
						"description": "Date of transaction ",
						"example": "Tue Nov 22 2022 12:58:30 GMT+0200"
					},
					"timestamps": {
						"type": "number",
						"description": "getTime() of transaction ",
						"example": "1669586400000"
					},

					"amount": {
						"type": "number",
						"description": "Sum of transaction ",
						"example": "100"
					},
					"balanceAfterTransaction": {
						"type": "string",
						"description": "Balance after transaction ",
						"example": "-50"
					},
					"owner": {
						"type": "string",
						"description": "Transaction's owner (id)",
						"example": "63837d687f0d6aafc551ffcc"
					},
					"createdAt": {
						"type": "string",
						"description": "Transaction's creating date",
						"example": "2022-11-27T15:08:24.572Z"
					}
				}
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}